requirements.txt
python-telegram-bot==21.6
openai>=1.35.0
httpx>=0.27.0
